+++ src/models/wordattention.py
@@ -32,8 +32,7 @@ class WordAttention(torch.nn.Module):
 
         assert self.recurrent_size % 2 == 0
 
-        if self.embedder:
-            assert self.word_emb_size == self.embedder.dim
+        assert self.word_emb_size == self.embedder.dim
 
         # embedding layer
         self.embedding = torch.nn.Embedding(num_embeddings=len(self.vocab), embedding_dim=self.word_emb_size)
